requirements.txt
# Core dependencies for news edit agent prototype
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Video processing
opencv-python==4.8.1.78
ffmpeg-python==0.2.0
pillow==10.1.0

# Audio transcription (MLX-optimized for M4 Mac)
mlx-whisper==0.3.0

# Embeddings and vector search
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3

# LLM client
anthropic==0.7.8
httpx==0.25.2

# Output formats
lxml==4.9.3

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# WebSocket support (for Open Arena Gemini)
websockets==12.0
certifi==2023.11.17
